user activity

+++ post-service/src/comment/comment.service.ts
@@ -1,30 +1,66 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { CommentDao } from './comment.dao';
-import { DeleteCommentInput } from './dtos/deleteCommentInput.dto';
-import { UpdateCommentInput } from './dtos/updateComment.dto';
 import { CommentResponseDto } from './dtos/commentResponse.dto';
 import { CreateCommentInput } from './dtos/createCommentInput.dto';
-
+import { DeleteCommentInput } from './dtos/deleteCommentInput.dto';
+import { UpdateCommentInput } from './dtos/updateComment.dto';
 
 @Injectable()
 export class CommentsService {
-  constructor(private readonly commentDao: CommentDao) {} // Inject `UserDao`
+  constructor (private readonly commentDao: CommentDao) {} // Inject `UserDao`
 
-  async createComment(input: CreateCommentInput) {
-    return this.commentDao.createCommentDao(input);
+  async createComment (input: CreateCommentInput) {
+    const dataObject = {
+      description: input.description,
+      createdBy: input.createdBy,
+      modifiedBy: input.createdBy,
+      postId: input.postId,
+      channelId: input.channelId,
+      createdAt: new Date(),
+      isDeleted: false,
+    };
+    return this.commentDao.createCommentDao(dataObject);
   }
 
- 
-
-  async getComments():Promise<CommentResponseDto[]> {
+  async getComments (): Promise<CommentResponseDto[]> {
     return this.commentDao.getCommentsDao();
   }
 
-  async deleteComment(input: DeleteCommentInput): Promise<string>{
-    return this.commentDao.deleteCommentDao(input)
+  async deleteComment (
+    input: DeleteCommentInput,
+    channelsAllowed: bigint[],
+    userId: bigint,
+    role: string
+  ): Promise<string> {
+    if (
+      await this.commentDao.canUserProceed(
+        input.id,
+        channelsAllowed,
+        userId,
+        role
+      )
+    ) {
+      return this.commentDao.deleteCommentDao(input);
+    }
+    throw new UnauthorizedException('User not allowed to delete this post');
   }
 
-  async updateComment(input: UpdateCommentInput): Promise<string>{
-    return this.commentDao.updateComment(input)
+  async updateComment (
+    input: UpdateCommentInput,
+    channelsAllowed: bigint[],
+    userId: bigint,
+    role: string
+  ): Promise<string> {
+    if (
+      await this.commentDao.canUserProceed(
+        input.id,
+        channelsAllowed,
+        userId,
+        role
+      )
+    ) {
+      return this.commentDao.updateComment(input);
+    }
+    throw new UnauthorizedException('User not allowed to delete this post');
   }
 }
