+++ post-service/src/reply/reply.service.ts
@@ -1,30 +1,67 @@
-import { Injectable } from '@nestjs/common';
-import { ReplyDao } from './reply.dao';
-import { UpdateReplyInput } from './dtos/updateReply.dto';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
+import { CreateReplyInput } from './dtos/createReply.dto';
 import { DeleteReplyInput } from './dtos/deleteReply.dto';
 import { ReplyResponseDto } from './dtos/replyComment.dto';
-import { CreateReplyInput } from './dtos/createReply.dto';
-
+import { UpdateReplyInput } from './dtos/updateReply.dto';
+import { ReplyDao } from './reply.dao';
 
 @Injectable()
 export class RepliesService {
-  constructor(private readonly replyDao: ReplyDao) {} // Inject `UserDao`
+  constructor (private readonly replyDao: ReplyDao) {} // Inject `UserDao`
 
-  async createReply(input: CreateReplyInput) {
-    return this.replyDao.createReplyDao(input);
+  async createReply (input: CreateReplyInput) {
+    const dataObject = {
+      description: input.description,
+      createdBy: input.createdBy,
+      modifiedBy: input.createdBy,
+      commentId: input.commentId,
+      postId: input.postId,
+      channelId: input.channelId,
+      createdAt: new Date(),
+      isDeleted: false,
+    };
+    return this.replyDao.createReplyDao(dataObject);
   }
 
- 
-
-  async getReplys():Promise<ReplyResponseDto[]> {
-    return this.replyDao.getReplysDao();
+  async getReplies (): Promise<ReplyResponseDto[]> {
+    return this.replyDao.getRepliesDao();
   }
 
-  async deleteReply(input: DeleteReplyInput): Promise<string>{
-    return this.replyDao.deleteReplyDao(input)
+  async deleteReply (
+    input: DeleteReplyInput,
+    channelsAllowed: bigint[],
+    userId: bigint,
+    role: string
+  ): Promise<string> {
+    if (
+      await this.replyDao.canUserProceed(
+        input.id,
+        channelsAllowed,
+        userId,
+        role
+      )
+    ) {
+      return this.replyDao.deleteReplyDao(input);
+    }
+    throw new UnauthorizedException('User not allowed to delete this post');
   }
 
-  async updateReply(input: UpdateReplyInput): Promise<string>{
-    return this.replyDao.updateReply(input)
+  async updateReply (
+    input: UpdateReplyInput,
+    channelsAllowed: bigint[],
+    userId: bigint,
+    role: string
+  ): Promise<string> {
+    if (
+      await this.replyDao.canUserProceed(
+        input.id,
+        channelsAllowed,
+        userId,
+        role
+      )
+    ) {
+      return this.replyDao.updateReply(input);
+    }
+    throw new UnauthorizedException('User not allowed to delete this post');
   }
 }
